+++ src/js/lib/card-nicknames/afr.ts
@@ -1,0 +1,59 @@
+// Adventures in the Forgotten Realms (AFR)
+export default [
+  {
+    // https://scryfall.com/card/afr/117/talisman-of-segador?utm_source=api
+    setCode: "afr",
+    collectorNumber: "117",
+    realName: ["Reaper's Talisman"],
+    nickname: ["Talisman of Segador"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/afr/126/moody-lord-of-the-night?utm_source=api
+    setCode: "afr",
+    collectorNumber: "126",
+    realName: ["Westgate Regent"],
+    nickname: ["Moody Lord of the Night"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/afr/379/hob-boss?utm_source=api
+    setCode: "afr",
+    collectorNumber: "147",
+    realName: ["Hobgoblin Bandit Lord"],
+    nickname: ["Hob Boss"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/afr/165/horny-bard?utm_source=api
+    setCode: "afr",
+    collectorNumber: "165",
+    realName: ["Valor Singer"],
+    nickname: ["Horny Bard"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/afr/192/venom-troll?utm_source=api
+    setCode: "afr",
+    collectorNumber: "192",
+    realName: ["Loathsome Troll"],
+    nickname: ["Venmo Troll"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/afr/234/targ-nar-large-beast-chaotic-evil?utm_source=api
+    setCode: "afr",
+    collectorNumber: "234",
+    realName: ["Targ Nar, Demon-Fang Gnoll"],
+    nickname: ["Targ Nar, Large Beast, Chaotic Evil"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/afr/259/hydra-nest?utm_source=api
+    setCode: "afr",
+    collectorNumber: "259",
+    realName: ["Lair of the Hydra"],
+    nickname: ["Hydra Nest"],
+    source: "Scryfall Preview Name",
+  },
+];